+++ otter/assign/output.py
@@ -7,6 +7,7 @@ import shutil
 import pathlib
 import warnings
 import nbformat
+import tempfile
 
 from .constants import NB_VERSION
 from .notebook_transformer import transform_notebook
@@ -36,14 +37,15 @@ def write_autograder_dir(nb_path, output_nb_path, assignment):
             warnings.warn("Could not auto-parse kernelspec from notebook; assuming Python")
             assignment.lang = "python"
 
-    output_dir = output_nb_path.parent
-    tests_dir = output_dir / 'tests'
-    os.makedirs(tests_dir, exist_ok=True)
-
     transformed_nb, test_files = transform_notebook(nb, assignment)
 
     # replace plugins
     transformed_nb = replace_plugins_with_calls(transformed_nb)
+
+    output_dir = output_nb_path.parent
+    tests_dir = output_dir / 'tests'
+    if assignment.test_files:
+        os.makedirs(tests_dir, exist_ok=True)
 
     if assignment.requirements:
         output_fn = ("requirements.txt", "requirements.R")[assignment.is_r]
@@ -61,15 +63,21 @@ def write_autograder_dir(nb_path, output_nb_path, assignment):
     # strip out ignored lines
     transformed_nb = strip_ignored_lines(transformed_nb)
 
-    # write notebook
-    # with open(output_nb_path) as f:
-    # nbformat.write(transformed_nb, )
-    nbformat.write(transformed_nb, str(output_nb_path))
-
     # write tests
     test_ext = (".R", ".py")[assignment.is_python]
     for test_name, test_file in test_files.items():
-        write_test(tests_dir / (test_name + test_ext), test_file)
+        test_path = tests_dir / (test_name + test_ext) if assignment.test_files else test_name
+        write_test(transformed_nb, test_path, test_file, use_file=assignment.test_files)
+
+    # write a temp dir for otter generate tests
+    if assignment.generate:
+        assignment._temp_test_dir = pathlib.Path(tempfile.mkdtemp())
+        for test_name, test_file in test_files.items():
+            test_path = assignment._temp_test_dir / (test_name + test_ext)
+            write_test(transformed_nb, test_path, test_file, use_file=True)
+
+    # write notebook
+    nbformat.write(transformed_nb, str(output_nb_path))
 
     # copy files
     for file in assignment.files:
@@ -126,11 +134,12 @@ def write_student_dir(nb_name, autograder_dir, student_dir, assignment):
 
     nb = strip_solutions_and_output(nb)
 
+    # remove hidden tests from student directory
+    remove_hidden_tests_from_dir(nb, student_dir / 'tests', assignment, use_files=assignment.test_files)
+
     with open(student_nb_path, "w") as f:
         nbformat.write(nb, f)
 
-    # remove hidden tests from student directory
-    remove_hidden_tests_from_dir(student_dir / 'tests', assignment)
 
 def write_output_directories(master_nb_path, result_dir, assignment):
     """