+++ otter/service/__init__.py
@@ -1,8 +1,0 @@
-"""
-Otter Service command-line utility. Provides support for managing a grading VM on a 3rd part cloud
-provider.
-"""
-
-from . import build
-from . import create
-from . import start