+++ docs/changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+**Unreleased:**
+
+* Added `NotebookMetadataTestFile` for reading tests from Jupyter Notebook metadata
+* Added optional storage of tests in notebook metadata for Otter Assign (default-off until v3)
+* Removed deprecated tool Otter Service
+* Added printout of total score to autograder output
+
 **v2.2.2:**
 
 * Added `show_all_public` configuration for grading